requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
opencv-python==4.9.0.80
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
numpy==1.26.4
aiofiles==23.2.1
